- allow swipe all over the screen - disable possibility game to be finished twice due to Finish listener rise condition

+++ app/src/main/java/com/tdanylchuk/app/engine/Game.kt
@@ -89,7 +89,7 @@ class Game(private val strategy: ContentStrategy) : Serializable {
 
     private fun createAnimator(): AnimatorSet {
         val animator = AnimatorSet()
-        animator.duration = GameplayConstants.ANIMATION_SPEED.toLong()
+        animator.duration = GameplayConstants.ANIMATION_SPEED
         animator.addListener(finishCheckListener)
         return animator
     }
